Rusken: - remove usage of sdk types - refactored controller.ts - moved code to helpers.ts and routingErrors.ts - added new dev params for testing - updated generated types
