+++ package.json
@@ -36,6 +36,7 @@
         "body-parser": "^1.20.0",
         "cors": "^2.8.5",
         "date-fns": "^2.28.0",
+        "clean-deep": "^3.4.0",
         "express": "^4.17.3",
         "express-jwt": "^6.1.1",
         "google-auth-library": "^8.0.3",
@@ -63,9 +64,9 @@
         "@types/lz-string": "^1.3.34",
         "@types/node": "16.11.6",
         "@types/node-fetch": "^2.6.1",
+        "@types/redis": "^2.8.32",
         "@types/semver": "^7.3.9",
         "@types/uuid": "^8.3.4",
-        "@types/redis": "^2.8.32",
         "@typescript-eslint/eslint-plugin": "^5.18.0",
         "@typescript-eslint/parser": "^5.18.0",
         "apigeetool": "^0.15.2",
